requirements.txt
streamlit
pandas
numpy
matplotlib.pyplot
scikit-learn
seaborn